+++ host/utilities/bladeRF-cli/src/cmd/flash_image.c
@@ -35,7 +35,7 @@
 #include "input.h"
 
 /* Previously BLADERF_FLASH_BYTE_LEN_FPGA. TODO: don't hardcode this */
-static size_t const LEGACY_FLASH_BYTE_LEN_FPGA = 0x00370000;
+static uint32_t const LEGACY_FLASH_BYTE_LEN_FPGA = 0x00370000;
 
 struct params
 {
